Move SessionConfig to datafusion_execution
